feat: BoardMapper 인터페이스 구현 및 id 자료형 Long으로 변경

+++ src/main/java/org/crud_ex/domain/Board.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class Board {
 
-	private int bno;
+	private Long bno;
 	private String title;
 	private String content;
 	private String writer;
